general: Use narrowing for error types
Errors in catch blocks can either be labeled `any` or `unknown`.
The semantics make a big difference. `any` here means
that a value can be anything, thus any method call or field
access is valid - essentially it turns off type checking for
the value. `unknown` however means that it's *something*,
but we need to figure out what.

We pair `unknown` with the `instanceof` operator in order
to narrow down the type to something specific. This allows
us to preserve TypeScript's type checking for error values
and their fields and methods.

See
https://www.typescriptlang.org/docs/handbook/2/narrowing.html#instanceof-narrowing