general: Use explicit public and private access modifiers
Previously we had a convention of using underscores to denote
that a field or function was private. However, this convention
was merely a soft guide that wasn't enforced by anything.
In practice this meant that some fields that should have been
marked as private weren't, and fields that were marked as private
were used outside of their intended files.

As part of the shift to TypeScript, which aims to use hard rules
to enforce conventions, we can adopt these access modifiers and
use them to clean up our code a bit. Where appropriate we've removed
the underscores on fields and functions and prepended access
modifiers appropriate for their use. The addition of these modifiers
allow us to know if a private field isn't in use, meaning that
we can clean up some dead fields.

We still use underscores in two scenarios:

* Fields generated by GObject.registerClass() use the underscore
convention. In order for us to align our TypeScript code we need
to continue using underscores for them when we declare them.
* Private fields that have public accessors still use underscores.