Fix the close_request vfunc implementation
GtkWindowClass.close_request() returns a boolean value:

- true, for stopping the signal emission
- false, for continuing the signal emission

The default behaviour when closing a window is to destroy it, which will
terminate the application instance.

The lack of a return value was masked by the explicit call to
GApplication.quit(), which forcibly terminates the application in the
following main loop iteration.