Fixed delete records and close
The record was being delete only when the toast was dismissed.

To ensure that the records are deleted I added a new variable undoToasts, that is
a Toast[] to have track of all the toasts that are being displayed.

When the windows is closed it dismiss all the toast remainings, causing all the
recordings that were temporarily deleted to be permanently deleted.