application: Fix directory creation
Creating the recordings dir can fail if the cache dir still exists,
as the existence of the cache dir will cause an error to be thrown
before the recordings dir is recreated.

This commit splits the logic of creation and error reporting
off into a separate function, allowing the directories to be
created independently.